Refactor seed task
