+++ seeds/003-add-kcal-flows.js
@@ -1,9 +1,65 @@
 const path = require("path");
 const fs = require("fs-extra");
-const { KCAL_FLOWS_DIR } = require("../config");
+const postgis = require("knex-postgis");
+const { KCAL_FLOWS_DIR, APP_DATA_DIR } = require("../config");
+
+// Generate absolute paths
+const CROPS_CSV_PATH = path.join(APP_DATA_DIR, "crops.csv");
+const COUNTIES_FILE_PATH = path.join(
+  APP_DATA_DIR,
+  "county-population-consumption-production-scaled.geojson"
+);
 
 exports.seed = async function (knex) {
-  // For each csv file in app-data/kcal-flows
+  /**
+   * CLEAR FLOWS, CROPS, AND COUNTIES TABLES
+   */
+  await knex.raw("TRUNCATE TABLE kcal_flows RESTART IDENTITY CASCADE");
+  await knex.raw("TRUNCATE TABLE crops RESTART IDENTITY CASCADE");
+  await knex.raw("TRUNCATE TABLE counties RESTART IDENTITY CASCADE");
+
+  /**
+   * INGEST COUNTIES
+   */
+  const { features: counties } = await fs.readJson(COUNTIES_FILE_PATH);
+  await knex.batchInsert(
+    "counties",
+    counties.map((county) => ({
+      id: county.properties.geography.slice(-5),
+      properties: county.properties,
+      geom: postgis(knex).geomFromGeoJSON(county.geometry),
+    })),
+    500
+  );
+
+  /**
+   * INGEST CROP CODES
+   */
+  const cropCodes = await fs.readFile(CROPS_CSV_PATH, "utf-8");
+
+  // Split CSV into rows
+  const rows = cropCodes.split("\n");
+
+  // Remove header row
+  rows.shift();
+
+  // Batch insert crop codes
+  await knex("crops").insert(
+    rows
+      .map((row) => {
+        const [id, name, category] = row.split(",");
+        return {
+          id,
+          name: name.trim(),
+          category,
+        };
+      })
+      .filter((row) => row.name !== "")
+  );
+
+  /**
+   * INGEST KCAL FLOWS
+   */
   const files = (await fs.readdir(KCAL_FLOWS_DIR)).filter((file) =>
     file.endsWith(".csv")
   );
@@ -34,7 +90,8 @@ exports.seed = async function (knex) {
 
     // Batch insert kcal flows
     try {
-      await knex("kcal_flows").insert(
+      await knex.batchInsert(
+        "kcal_flows",
         rows
           .map((row) => {
             const [origin_id, destination_id, value] = row.split(",");
@@ -51,7 +108,8 @@ exports.seed = async function (knex) {
               value,
             };
           })
-          .filter((row) => row.origin_id !== "Dade")
+          .filter((row) => row.origin_id !== "Dade"),
+        500
       );
     } catch (error) {
       console.log(error);
